% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{as.genalex}
\alias{as.genalex}
\title{Convert pre-1.0 readGenalex data frames to class 'genalex'}
\usage{
as.genalex(x)
}
\arguments{
\item{x}{An annotated data frame that is of \code{readGenalex} format,
          possibly of the earlier pre-1.0 format}
}
\value{
\code{x} as an S3 class \code{'genalex'} object, or
        a fatal error
}
\description{
Converts an annotated data frame that is of \code{readGenalex} format
to a class \code{'genalex'} object.  If the data frame is a pre-1.0
data frame, it is converted to class \code{'genalex'}, if it is already
of class \code{'genalex'} object it is returned, and if it is of any
other format, an error is generated.
}
\examples{
data(example_genotypes)
as.genalex(example_genotypes)
}
\author{
Douglas G. Scofield
}

