% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{cbind.genalex}
\alias{cbind.genalex}
\title{Combine loci from class \code{'genalex'} data sets}
\usage{
cbind.genalex(..., names, deparse.level = 1)
}
\arguments{
\item{names}{List of names: \code{title} for data set title,
\code{sample} for sample column header, and \code{pop} for population
column header.  If \code{names} or any of its fields are not provided,
the names of the first argument to \code{\dots} are used.}

\item{deparse.level}{Not used (yet)}

\item{\dots}{All arguments must be class \code{'genalex'} data sets.
If only one data set is supplied, it is returned unmodified.}
}
\value{
Annotated data frame of class \code{'genalex'}.  If \code{names}
or any of its fields are not provided, the names of the first argument
are used.  The \code{data.file.name} attribute is a character
representation of the call to \code{cbind}.
}
\description{
Combine loci from class \code{'genalex'} data sets into one larger class
\code{'genalex'} data set by adding loci.  Locus counts are adjusted
accordingly.  The data sets must have the same ploidy, sample names
and sample membership in populations.  The order of the samples may
differ, and the final data set will have the sample order of the first.
Locus names must be unique across all data sets; if any locus names are
duplicated, the genotypes are checked to assure they contain the exact
same data and if so, the duplicate locus columns are ignored.  Data set
title and sample and population column headers are taken from the first
data set unless supplied in the \code{names} argument.  Extra columns are
taken from the first data set; if additional data sets have extra columns,
their contents must match for columns with the same name, and columns
with different names are added to the extra columns of the returned
data set.
}
\details{
Data sets must have the same number of rows, this is a necessary
consequence of containing the same samples.  Unlike the default
\code{cbind} method, values will not be recycled to create matching
row counts.
}
\note{
If one of the arguments is class \code{'data.frame'}, then this
function will \emph{not} be called, instead the \code{cbind.data.frame}
method of base R will be called silently and will return an object of
class \code{'data.frame'}.  If this happens, none of the special
processing that selects just locus columns and combines extra columns
from class \code{'genalex'} objects occurs, so if you have mixed class
\code{'genalex'} objects with data frames the return value is probably
not what you intended.  This call to \code{cbind.data.frame} occurs
because objects of class \code{'genalex'} also have class
\code{'data.frame'}, and selection of \code{cbind.data.frame} occurs
during method dispatch for \code{rbind} and cannot be checked by this
function.  If there is a chance you have mixed objects of different
classes while calling this function, assure that the return value is
class \code{'genalex'} by using \code{is.genalex}.
}
\author{
Douglas G. Scofield
}
\seealso{
\code{\link{genalex}}, \code{\link{cbind}}, \code{\link{cbind.genalex}}
}

