% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkNullAlleles.R
\name{checkNullAlleles}
\alias{checkNullAlleles}
\title{Check for potential null (non-amplifying) alleles}
\usage{
checkNullAlleles(refgt, comparegt, quiet = FALSE)
}
\arguments{
\item{quiet}{\code{TRUE} or \code{FALSE}, if \code{TRUE} then informational
              messages are printed during the check for null alleles}
}
\value{
Nothing.  Informational messages are printed during its execution.
        Perhaps this should be changed so that it returns a list of
        coordinates within the genotypes that represent potential null
        alleles?
}
\description{
Compare genotypes against a set of reference genotypes and look for
genotypes that differ in a way consistent with the occurrence of a null
allele, that is, heterozygous in the reference and homozygous matching one
of the alleles in the heterozygous reference.
}
\details{
The reference and compare genotypes are in GenAlEx format, and are either
names of files to be read as GenAlEx files, or are class \code{'genalex'}
data frames.
}
\examples{
# load reference genotypes
# load comparison genotypes
# run comparison
}
\author{
Douglas G. Scofield
}

