% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{computeGenalexColumns}
\alias{computeGenalexColumns}
\title{Determine numeric column positions occupied by named loci}
\usage{
computeGenalexColumns(dat, locus, ploidy = NULL)
}
\arguments{
\item{dat}{An annotated data frame created by \code{readGenalex}}

\item{locus}{The names of one or more loci found in \code{dat}}

\item{ploidy}{Ploidy of data in \code{dat}, if not supplied is extracted
              from the \code{ploidy} attribute of \code{dat}}
}
\value{
A vector of column positions occupied by genotype data for loci
        named in \code{locus}.
}
\description{
Determine the numeric column positions occupied by named loci in a
date frame produced by \code{readGenalex}.  This is mostly used as
a utility routine by other functions in the \code{readGenalex} package.
}
\examples{
data(example_genotypes)
computeGenalexColumns(example_genotypes, c("loc2","loc4"))
}
\author{
Douglas G. Scofield
}

