% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{dropGenalexLoci}
\alias{dropGenalexLoci}
\title{Remove specified loci from data frame of class \code{'genalex'}}
\usage{
dropGenalexLoci(x, drop.loci, quiet = FALSE)
}
\arguments{
\item{x}{An annotated data frame of class \code{'genalex'}}

\item{drop.loci}{The names of one or more loci found in \code{x}}

\item{quiet}{If set to \code{TRUE}, quietly returns \code{x}
                 if none of \code{drop.loci} are found in \code{x}}
}
\value{
A data frame containing the data in \code{x} after removing
loci specified by \code{drop.loci}, with attributes updated as required.
}
\description{
Remove specified loci from the data frame of class \code{'genalex'}
and updates attributes
}
\examples{
data(Qagr_adult_genotypes)
newdat <- dropGenalexLoci(Qagr_adult_genotypes, "Oe09")
}
\author{
Douglas G. Scofield
}

