% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{dropLocus}
\alias{dropLocus}
\alias{dropLocus.genalex}
\title{Remove specified loci from data frame of class \code{'genalex'}}
\usage{
\method{dropLocus}{genalex}(x, drop.locus, quiet = FALSE, ...)
}
\arguments{
\item{x}{An annotated data frame of class \code{'genalex'}}

\item{drop.locus}{The names of one or more loci found in \code{x}}

\item{quiet}{If set to \code{TRUE}, quietly returns \code{x}
                  if none of \code{drop.locus} are found in \code{x}}

\item{\dots}{Additional arguments, currently none}
}
\value{
A data frame containing the data in \code{x} after removing
loci specified by \code{drop.locus}, with attributes updated as
required.
}
\description{
Remove specified loci from the data frame of class \code{'genalex'}
and updates attributes
}
\examples{
data(Qagr_adult_genotypes)
newdat <- dropLocus(Qagr_adult_genotypes, "Oe09")
}
\author{
Douglas G. Scofield
}

