% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{is.genalex}
\alias{is.genalex}
\title{Check to see if an object is of class 'genalex'}
\usage{
is.genalex(x, force = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{An object that might be of class \code{'genalex'}}

\item{force}{If \code{TRUE} and \code{x} has class \code{'genalex'},
force a deeper check to assure that the data and annotations are
consistent with class \code{'genalex'}}

\item{verbose}{If \code{TRUE} and \code{force = TRUE}, indicate any
inconsistencies discovered between the data and annotations. if
there are no inconsistencies, nothing is printed.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check to see if an object is of class 'genalex' as recognised by
the \code{readGenalex} package.  With \code{force = TRUE},
the internal consistency of the data and annotations are checked.
}
\examples{
data(Qagr_adult_genotypes)
is.genalex(Qagr_adult_genotypes)
## create an inconsistent attribute
attr(Qagr_adult_genotypes, "n.loci") <- 20
## this doesn't detect the inconsistency
is.genalex(Qagr_adult_genotypes)
## this detects the inconsistency, returns FALSE
is.genalex(Qagr_adult_genotypes, force = TRUE)
is.genalex(Qagr_adult_genotypes, force = TRUE, verbose = TRUE)
}
\author{
Douglas G. Scofield
}

