% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex-pegas.R
\name{joinGenotypes}
\alias{joinGenotypes}
\alias{joinGenotypes.default}
\alias{joinGenotypes.genalex}
\title{Join genotypes encoded as separate alleles into single genotypes like class \code{'genotype'} or class \code{'loci'}}
\usage{
\method{joinGenotypes}{genalex}(x, sep = "/", ...)

\method{joinGenotypes}{default}(x, loci, ploidy, sep = "/", ...)
}
\arguments{
\item{x}{An object of class \code{'genalex'}, a data frame, or an
object that can be coerced to a data frame which contains genotypes
with alleles in separate columns, to be joined into character
genotypes of the form \code{"101/107"}, where the \code{"/"} is
determined by \code{sep}.}

\item{sep}{Separator between alleles}

\item{loci}{Columns or column names indicating loci to join.  When
called for an object of class \code{'genalex'}, all loci are joined.}

\item{ploidy}{Ploidy of the indicated loci.  When called for an
object of class \code{'genalex'}, ploidy is determined from attributes.}

\item{\dots}{Additional arguments, currently ignored}
}
\value{
A data frame with allele columns replaced by genotype columns
}
\description{
Join genotypes encoded as separate alleles into single genotypes like class \code{'genotype'} or class \code{'loci'}
}

