% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{ploidy}
\alias{ploidy}
\alias{ploidy.genalex}
\title{Return ploidy for object of class \code{'genalex'}}
\usage{
\method{ploidy}{genalex}(x, ...)
}
\arguments{
\item{x}{An annotated data frame of class \code{'genalex'}}

\item{\dots}{Additional arguments, currently ignored}
}
\value{
The numeric value of the \code{"ploidy"} attribute of \code{x}
}
\description{
Return ploidy for object of class \code{'genalex'}
}
\examples{
data(Qagr_pericarp_genotypes)
ploidy(Qagr_pericarp_genotypes)
}
\author{
Douglas G. Scofield
}

