% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex-data.R
\docType{data}
\name{Qagr_adult_genotypes}
\alias{Qagr_adult_genotypes}
\title{Coast live oak (Quercus agrifolia) adult microsatellite genotypes}
\format{Annotated data frames of class \code{'genalex'}.  Columns are:
\tabular{ll}{
  \code{AdultMomFamily} \tab Identifier for adult tree \cr
  \code{Site}           \tab Population, always \code{Sedgwick} for this data set \cr
  \code{0c11}           \tab First allele for microsatellite locus 0c11 \cr
  \code{0c11.2}         \tab Second allele for microsatellite locus 0c11 \cr
  \code{\dots}          \tab Remaining 9 microsatellite genotypes \cr
}}
\source{
\url{http://datadryad.org/resource/doi:10.5061/dryad.40kq7}
}
\description{
This data set contains one annotated data frame of class \code{'genalex'},
holding 10-locus diploid microsatellite genotypes of 262 adult coast live
oak (\emph{Quercus agrifolia}) trees from Sedgwick Reserve, Santa Barbara
County, California, USA.
}
\details{
These data have been analysed in several published papers, and are also
publicly available at the Dryad Digital Repository.  If using these data,
please cite the original paper as well as the data, as shown below.
}
\references{
Scofield DG, Smouse PE, Karubian J, Sork VL (2012) Use of alpha,
beta, and gamma diversity measures to characterize seed dispersal by
animals. The American Naturalist 180(6): 719-732.
\url{http://dx.doi.org/10.1086/668202}

Scofield DG, Smouse PE, Karubian J, Sork VL (2012) Data from: Use of alpha,
beta, and gamma diversity measures to characterize seed dispersal by animals.
Dryad Digital Repository. \url{http://dx.doi.org/10.5061/dryad.40kq7}

\url{https://github.com/douglasgscofield/readGenalex}
}
\keyword{datasets}

