% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\docType{data}
\name{Qagr_genotypes}
\alias{Qagr_genotypes}
\title{Quercus agrifolia adult and pericarp microsatellite genotypes}
\format{Annotated data frames of class \code{'genalex'}}
\source{
\url{http://datadryad.org/resource/doi:10.5061/dryad.40kq7}
}
\description{
This data set contains two annotated data frames of class \code{'genalex'}.
}
\details{
\itemize{
  \item \code{Qagr_adult_genotypes} contains 10-locus diploid microsatellite
  genotypes of 262 adult coast live oak (\emph{Quercus agrifolia}) trees from
  Sedgwick Reserve, Santa Barbara County, California, USA.

\item \code{Qagr_peric_genotypes} contains 10-locus diploid microsatellite
  genotypes of 568 pericarps (outer seed coats) from coast live oak acorns
  collected from 17 acorn woodpecker (\emph{Melanerpes formicivorus})
  granaries, also at Sedgwick Reserve.
}

These data have been analysed in several published papers, and are also
publicly available at the Dryad Digital Repository.  If using these data,
please cite the original paper as well as the data, as shown below.
}
\references{
Scofield DG, Smouse PE, Karubian J, Sork VL (2012) Use of alpha,
beta, and gamma diversity measures to characterize seed dispersal by
animals. The American Naturalist 180(6): 719-732.
\url{http://dx.doi.org/10.1086/668202}

Scofield DG, Smouse PE, Karubian J, Sork VL (2012) Data from: Use of alpha,
beta, and gamma diversity measures to characterize seed dispersal by animals.
Dryad Digital Repository. \url{http://dx.doi.org/10.5061/dryad.40kq7}

\url{https://github.com/douglasgscofield/readGenalex}
}
\keyword{datasets}

