% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex-deprecated.R
\docType{package}
\name{readGenalex-deprecated}
\alias{computeGenalexColumns}
\alias{dropGenalexLoci}
\alias{getGenalexLocus}
\alias{printGenalexGenotype}
\alias{putGenalexLocus}
\alias{readGenalex-deprecated}
\alias{readGenalex-deprecated-package}
\alias{reduceGenalexPloidy}
\alias{reorderGenalexLoci}
\title{Deprecated functions in the \code{readGenalex} package}
\usage{
printGenalexGenotype(...)
}
\arguments{
\item{...}{Parameters to be passed to the current version of the function}
}
\description{
These functions are provided for compatibility with older version of
the \code{readGenalex} package.  They may eventually be completely removed.
}
\section{Details}{


\tabular{rl}{
  \code{printGenalexGenotype} \tab now a synonym for \code{\link{printGenotype}}\cr
  \code{computeGenalexColumns} \tab now a synonym for \code{\link{getLocusColumns}}\cr
  \code{reorderGenalexLoci} \tab now a synonym for \code{\link{reorderLoci}}\cr
  \code{getGenalexLocus} \tab now a synonym for \code{\link{getLocus}}\cr
  \code{putGenalexLocus} \tab now a synonym for \code{\link{replaceLocus}}\cr
  \code{dropGenalexLoci} \tab now a synonym for \code{\link{dropLocus}}\cr
  \code{reduceGenalexPloidy} \tab now a synonym for \code{\link{reducePloidy}}\cr
}
}

