% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{reorderGenalexLoci}
\alias{reorderGenalexLoci}
\title{Reorder a data frame of class \code{'genalex'}}
\usage{
reorderGenalexLoci(x, loci)
}
\arguments{
\item{x}{An annotated data frame of class \code{'genalex'}}

\item{loci}{The names of loci found in \code{x}, in the desired new
            order.  All loci in \code{x} must be named.  The order of
            the alleles within each locus is preserved.}
}
\value{
A data frame of class \code{'genalex'} containing the same
genotype data from \code{x} reordered according to \code{loci}.
}
\description{
Reorder the genotype columns of a class \code{'genalex'} data frame by
locus.
}
\examples{
data(Qagr_adult_genotypes)
# reverse loci
loci <- rev(attr(Qagr_adult_genotypes, "locus.names"))
reord = reorderGenalexLoci(Qagr_adult_genotypes, rev(loci))
}
\author{
Douglas G. Scofield
}

