% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{replaceLocus}
\alias{replaceLocus}
\alias{replaceLocus.genalex}
\title{Replace genotype data in data frame of class \code{'genalex'}}
\usage{
\method{replaceLocus}{genalex}(x, locus, newdata, ...)
}
\arguments{
\item{x}{A data frame of class \code{'genalex'}}

\item{locus}{The names of one or more loci found in \code{x}}

\item{newdata}{New genotype data for loci specified in \code{locus}.  Must
               have the same number of rows as \code{x}}

\item{\dots}{Additional arguments, currently ignored}
}
\value{
A data frame of class \code{'genalex'} containing genotype data
from \code{x} with data for loci specified in \code{locus} replaced
with data from \code{newdata}.
}
\description{
Replace genotype data for specified loci in a data frame of class
\code{'genalex'}.
}
\author{
Douglas G. Scofield
}

