% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex-genetics.R
\name{splitGenotypes}
\alias{splitGenotypes}
\title{Split genotypes encoded like class 'genotype' into separate alleles}
\usage{
splitGenotypes(x, sep = "/")
}
\arguments{
\item{x}{A data frame or object that can be coerced to a data frame
which contains genotypes of the form \code{101/107}, where the
\code{"/"} is determined by \code{sep}}

\item{sep}{Separator between alleles}
}
\description{
Split genotypes encoded like class 'genotype' into separate alleles
}

