% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{summary.genalex}
\alias{summary.genalex}
\title{Summarise contents of class 'genalex' data frame}
\usage{
\method{summary}{genalex}(object, ...)
}
\arguments{
\item{object}{An annotated data frame of class \code{'genalex'}}

\item{...}{Additional arguments passed to \code{'summary'}
               for the data frame}
}
\value{
Result of \code{summary.data.frame(object)}
}
\description{
This prints a few lines summarising the data set title, sample size,
population sizes, ploidy, number of loci, and locus names, followed
by a summary of the data frame and a summary of the extra columns,
if present.
}
\examples{
data(Qagr_adult_genotypes)
summary(Qagr_adult_genotypes)
}
\author{
Douglas G. Scofield
}

