% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{writeGenalex}
\alias{writeGenalex}
\title{Write GenAlEx-format genotypes to a text file}
\usage{
writeGenalex(x, file, sep = "\\t", na = "", eol = "\\n")
}
\arguments{
\item{x}{Annotated \code{data.frame} created via the \code{readGenalex}
             package.}

\item{file}{File name or connection for writing.  If given as \code{""},
             \code{stdout()} is used.}

\item{sep}{Column separator for output (defaults to \code{"\t"}).}

\item{na}{The string to use when writing missing values in the data
             (defaults to \code{""}).}

\item{eol}{End-of-line character used for output (defaults to \code{"\n"}).}
}
\value{
No value is returned.
}
\description{
Writes genotype data encoded in an annotated \code{data.frame} via the
\code{readGenalex} package to a GenAlEx-format text file.  Extra data
columns are included immediately to the right of genotype columns.  GenAlEx
and its documentation are available at
\url{http://biology-assets.anu.edu.au/GenAlEx}.
}
\details{
This function writes genotypes and associated information within an
annotated \code{data.frame} to a text file in GenAlEx format. More
information is available in the description for \code{\link{readGenalex}},
and at the GenAlEx website at
\url{http://biology-assets.anu.edu.au/GenAlEx}.

Doing \code{writeGenalex(readGenalex("file.txt"), "file-write.txt")} won't
necessarily produce an output file identical to the input file.  Two areas
for which this will likely be true are:

\enumerate{
   \item Names on columns for alleles other than the first in a locus,
         which are ignored by \code{readGenalex}, converted to a simple
         concatenation of locus name and allele number in the resulting
         \code{data.frame}, and are left out of the output of
         \code{writeGenalex}.
   \item Locations of additional data columns beyond the genotype columns,
         which \code{readGenalex} will collect wherever there are named
         columns to the right of the genotype columns, and which
         \code{writeGenalex} will write immediately to the right of the
         genotype columns.  The same column names are used when writing
         as were present when reading.
}
}
\examples{
data(example_genotypes)
writeGenalex(example_genotypes, file = "")
}
\author{
Douglas G. Scofield
}
\references{
Peakall, R. and Smouse P.E. (2012) GenAlEx 6.5: genetic analysis
in Excel. Population genetic software for teaching and research-an update.
\emph{Bioinformatics} 28, 2537-2539.

Peakall, R. and Smouse P.E. (2006) GENALEX 6: genetic analysis in Excel.
Population genetic software for teaching and research. \emph{Molecular
Ecology Notes} 6, 288-295.
}
\seealso{
\code{\link{readGenalex}}
}
\keyword{attribute}
\keyword{file}
\keyword{manip}

