% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{as.genalex}
\alias{as.genalex}
\alias{as.genalex.data.frame}
\alias{as.genalex.genalex}
\title{Convert object to class 'genalex'}
\usage{
\method{as.genalex}{genalex}(x, ...)

\method{as.genalex}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object}

\item{\dots}{Additional arguments, currently ignored}
}
\value{
\code{x} as a class \code{'genalex'} object
}
\description{
Converts object \code{x} to a data frame of class \code{'genalex'}.
There are three cases:
\itemize{
  \item If \code{x} is of class \code{'genalex'}, it is simply returned.
  \item If \code{x} is of class \code{'data.frame'}, it is examined to
        see if it might be a data frame created by an earlier version of
        the \code{readGenalex} package.  If so, it is converted to
        class \code{'genalex'} and returned.
  \item Any other class is an error.  Further conversions between genetic
        data formats may be added here as additional methods.
}
}
\examples{
data(Qagr_adult_genotypes)
gt <- as.genalex(Qagr_adult_genotypes)
}
\author{
Douglas G. Scofield
}

