% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{as.genalex}
\alias{as.genalex}
\title{Convert pre-1.0 readGenalex data.frames to class 'genalex'}
\usage{
as.genalex(x)
}
\arguments{
\item{x}{An annotated \code{data.frame} that is a pre-1.0
          readGenalex data frame}
}
\value{
\code{x} as an S3 class \code{'genalex'} object, or
        a fatal error
}
\description{
Convert pre-1.0 readGenalex data.frames to class 'genalex'
}
\examples{
data(example_genotypes)
as.genalex(example_genotypes)
}
\author{
Douglas G. Scofield
}

