% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{as.genalex}
\alias{as.genalex}
\title{Convert pre-1.0 readGenalex data frames to class 'genalex'}
\usage{
as.genalex(x)
}
\arguments{
\item{x}{An object to convert to class \code{'genalex'}, possibly
          of an earlier package \code{readGenalex} format}
}
\value{
\code{x} as an S3 class \code{'genalex'} object, or
        a fatal error
}
\description{
Converts a data frame to a class \code{'genalex'} object.  The
following cases are handled:
\itemize{
  \item If the data frame is a pre-1.0 data frame, it is converted
  to class \code{'genalex'}
  \item If the data frame is already of class \code{'genalex'}, it
  is returned
  \item If \code{x} is of any other format, it is an error
}
}
\examples{
data(Qagr_adult_genotypes)
gt <- as.genalex(Qagr_adult_genotypes)
}
\author{
Douglas G. Scofield
}

