% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\docType{data}
\name{example_genotypes}
\alias{example_genotypes}
\title{Example genotypes for demonstrating readGenalex package functions}
\format{Data frame of class \code{'genalex'} containing 16 samples
        from two populations, with each sample having a completely
        hypothetical 5-locus genotype.}
\source{
Simulation
}
\description{
This data set contains 16 hypothetical genotypes used for demonstrating the
operation of various functions in the \code{readGenalex} package.  It is a
data frame of class \code{'genalex'} which contains additional
attributes describing the structure of the data set.
}
\keyword{datasets}

