\name{example_genotypes}
\docType{data}
\alias{example_genotypes}
\title{Example genotypes for demonstrating readGenalex package functions}
\description{
  This data set contains 16 hypothetical genotypes used for
  demonstrating the operation of various functions in the
  \code{readGenalex} package.  It is a \code{data.frame} created
  by \code{readGenalex()} and has additional attributes describing
  the structure of the data set.
}
\usage{example_genotypes}
\format{A \code{data.frame} containing 16 samples from two populations, with each sample having a completely hypothetical 5-locus genotype.}
\keyword{datasets}
