% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{is.genalex}
\alias{is.genalex}
\title{Check to see if a data frame is of class 'genalex'}
\usage{
is.genalex(x)
}
\arguments{
\item{x}{An data frame that might also be of class \code{'genalex'}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check to see if a data frame is of class 'genalex' as recognised by
the \code{readGenalex} package.
}
\examples{
data(example_genotypes)
is.genalex(example_genotypes)
}
\author{
Douglas G. Scofield
}

