% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{printGenalexGenotype}
\alias{printGenalexGenotype}
\title{Print selected genotypes}
\usage{
printGenalexGenotype(dat, rows, callout.locus = NULL, sep = " ",
  allele.sep = "/", callout.char = "*", label = NULL)
}
\arguments{
\item{dat}{An annotated \code{data.frame} created by \code{readGenalex()}}

\item{rows}{The specific rows of \code{dat} to print}

\item{callout.locus}{One or more loci on \code{dat} to be surrounded by
\code{callout.char} when printed}

\item{sep}{Separator character to be used between loci}

\item{allele.sep}{Separator character to be used between alleles}

\item{callout.char}{Character which surrounds loci specified by
\code{callout.locus}}

\item{label}{Label to be included between the sample and population ID
        columns and the genotype columns in output}
}
\value{
No specific return value, used for its side effect of printing
        genotypes.
}
\description{
Print selected genotypes, optionally calling out a locus
}
\examples{
data(example_genotypes)
printGenalexGenotype(example_genotypes, rows=6:8, callout.locus="loc5")
}
\author{
Douglas G. Scofield
}

