% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{putGenalexLocus}
\alias{putGenalexLocus}
\title{Replace genotype data in a \code{readGenalex}-format \code{data.frame}}
\usage{
putGenalexLocus(dat, locus, newdata)
}
\arguments{
\item{dat}{An annotated \code{data.frame} created by \code{readGenalex()}}

\item{locus}{The names of one or more loci found in \code{dat}}

\item{newdata}{New genotype data for loci specified in \code{locus}.  Must
have the same number of rows as \code{dat}}
}
\value{
A \code{data.frame} containing genotype data from \code{dat} with
data for loci specified in \code{locus} replaced with data from
\code{newdata}.
}
\description{
Replace genotype data for specified loci in a \code{readGenalex}-format
\code{data.frame}.
}
\author{
Douglas G. Scofield
}

