% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{reorderGenalexLoci}
\alias{reorderGenalexLoci}
\title{Reorder a data frame of class \code{'genalex'}}
\usage{
reorderGenalexLoci(dat, loci)
}
\arguments{
\item{dat}{An annotated data frame of class \code{'genalex'}}

\item{loci}{The names of loci found in \code{dat}, in the desired new
            order.  All loci in \code{dat} must be named.  The order of
            the alleles within each locus is preserved.}
}
\value{
A data frame of class \code{'genalex'} containing the same
genotype data from \code{dat} reordered according to \code{loci}.
}
\description{
Reorder the genotype columns of a class \code{'genalex'} data frame by
locus.
}
\examples{
data(example_genotypes)
# reverse loci
reord = reorderGenalexLoci(example_genotypes, rev(attr(example_genotypes, "locus.names")))
}
\author{
Douglas G. Scofield
}

