% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readGenalex.R
\name{reorderGenalexLoci}
\alias{reorderGenalexLoci}
\title{Reorder a \code{readGenalex}-format \code{data.frame}}
\usage{
reorderGenalexLoci(dat, loci)
}
\arguments{
\item{dat}{An annotated \code{data.frame} created by \code{readGenalex()}}

\item{loci}{The names of loci found in \code{dat}, in the desired new order.
All loci in \code{dat} must be named.  The order of the alleles within each
locus is preserved.}
}
\value{
A \code{data.frame} containing the same genotype data from
\code{dat} reordered according to \code{loci}.
}
\description{
Reorder the genotype columns of a \code{readGenalex}-format
\code{data.frame} by locus.
}
\examples{
data(example_genotypes)
# reverse loci
reord = reorderGenalexLoci(example_genotypes, rev(attr(example_genotypes, "locus.names")))
}
\author{
Douglas G. Scofield
}

